hps_accel/gateware: move math ops to post_process
Move SaturatingRoundingDoubleHighMul and RoundingDivideByPOT from
math.py to post_process.py to avoid a circular dependency.

In future, MathInstruction could possibly be renamed and moved to
post_process.py.
